# backend/api-server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SLIME Tools External API",
    "version": "1.0.0",
    "description": "External API for third-party project integrations. All endpoints are scoped to programs created by the API key holder. Contact SLIME Tools to request an API key."
  },
  "servers": [
    { "url": "https://api.slime.tools/api/v1/external" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/staking-programs/{id}": {
      "get": {
        "summary": "Get staking program details + stats",
        "operationId": "getStakingProgram",
        "description": "Returns details for a single program. Must be owned by the API key holder.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Program details with participant count and total distributed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "program": { "type": "object" },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "participantCount": { "type": "integer" },
                        "totalDistributed": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized for this program" },
          "404": { "description": "Program not found" }
        }
      }
    },
    "/staking-programs/{id}/position/{accountId}": {
      "get": {
        "summary": "Get a wallet's staking position",
        "operationId": "getStakingPosition",
        "description": "Live snapshot of a wallet's staking position in one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "accountId", "in": "path", "required": true, "schema": { "type": "string", "example": "0.0.12345" } }
        ],
        "responses": {
          "200": {
            "description": "Wallet staking position",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "accountId": { "type": "string" },
                    "programId": { "type": "string" },
                    "isRegistered": { "type": "boolean" },
                    "holdings": {
                      "type": "object",
                      "properties": {
                        "unitsHeld": { "type": "number" },
                        "meetsMinimum": { "type": "boolean" }
                      }
                    },
                    "totalEarned": { "type": "number" },
                    "nextDripAt": { "type": "string", "format": "date-time" },
                    "estimatedNextReward": { "type": "number" }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized for this program" }
        }
      }
    },
    "/staking-programs/{id}/eligibility/{accountId}": {
      "get": {
        "summary": "Check eligibility without registering",
        "operationId": "checkEligibility",
        "description": "Lightweight eligibility check for a wallet in one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "accountId", "in": "path", "required": true, "schema": { "type": "string", "example": "0.0.12345" } }
        ],
        "responses": {
          "200": {
            "description": "Eligibility result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "isRegistered": { "type": "boolean" },
                    "isEligible": { "type": "boolean" },
                    "holdings": { "type": "number" },
                    "estimatedReward": { "type": "number" },
                    "currency": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized for this program" }
        }
      }
    },
    "/staking-programs/{id}/register": {
      "post": {
        "summary": "Register a wallet and trigger immediate drip",
        "operationId": "registerParticipant",
        "description": "Registers a wallet and triggers an immediate reward drip in one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": { "type": "string", "example": "0.0.12345" }
                },
                "required": ["accountId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration result" },
          "400": { "description": "Not eligible or already registered" },
          "403": { "description": "Not authorized for this program" },
          "404": { "description": "Program not found" }
        }
      }
    },
    "/staking-programs/{id}/participants": {
      "get": {
        "summary": "List participants",
        "operationId": "listParticipants",
        "description": "Returns all participants for one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Participant list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "participants": { "type": "array" },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized for this program" }
        }
      }
    },
    "/staking-programs/{id}/distributions": {
      "get": {
        "summary": "List distribution history",
        "operationId": "listDistributions",
        "description": "Returns reward distribution history for one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Distribution history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "distributions": { "type": "array" }
                  }
                }
              }
            }
          },
          "403": { "description": "Not authorized for this program" }
        }
      }
    }
  }
}